requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_pytorch_select=0.2=gpu_0
argon2-cffi=20.1.0=py37h7b6447c_1
async_generator=1.10=py37h28b3542_0
attrs=20.3.0=pyhd3eb1b0_0
backcall=0.2.0=py_0
blas=1.0=mkl
bleach=3.2.1=py_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2020.10.14=0
cairo=1.14.12=h8948797_3
certifi=2020.11.8=py37h06a4308_0
cffi=1.14.3=py37h261ae71_2
chardet=3.0.4=py37h06a4308_1003
cloudpickle=1.6.0=py_0
cryptography=3.2.1=py37h3c74f83_1
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cycler=0.10.0=py37_0
cytoolz=0.11.0=py37h7b6447c_0
dask-core=2.30.0=py_0
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py37_0
expat=2.2.10=he6710b0_2
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.0.0=hf484d3e_5
freetype=2.10.4=h5ab3b9f_0
glib=2.66.1=h92f7085_0
graphite2=1.3.14=h23475e2_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
h5py=2.8.0=py37h989c5e5_3
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=he6710b0_3
idna=2.8=py37_0
imageio=2.9.0=py_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
iniconfig=1.1.1=py_0
intel-openmp=2019.4=243
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.13.0=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.5.1=py_1
jasper=2.0.14=h07fcdf6_1
jedi=0.15.1=py37_0
jinja2=2.11.2=py_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=py_2
jupyter=1.0.0=py37_7
jupyter_client=6.1.7=py_0
jupyter_console=6.2.0=py_0
jupyter_core=4.7.0=py37h06a4308_0
jupyterlab_pygments=0.1.2=py_0
kiwisolver=1.3.0=py37h2531618_0
lcms2=2.11=h396b838_0
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libvpx=1.7.0=h439df22_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.2=heb0550a_3
markupsafe=1.1.1=py37h14c3975_1
matplotlib=3.1.1=py37h5429711_0
mistune=0.8.4=py37h14c3975_1001
mkl=2019.4=243
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.0.2=py37hd81dba3_0
more-itertools=8.6.0=pyhd3eb1b0_0
nbclient=0.5.1=py_0
nbconvert=6.0.7=py37_0
nbformat=5.0.8=py_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.4.3=pyhd3eb1b0_0
networkx=2.5=py_0
ninja=1.10.1=py37hfd86e86_0
notebook=6.1.4=py37_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
openssl=1.1.1h=h7b6447c_0
packaging=20.4=py_0
pandas=0.25.1=py37he6710b0_0
pandoc=2.11=hb0f4dca_0
pandocfilters=1.4.3=py37h06a4308_1
parso=0.8.0=py_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=py37_1001
pillow=8.0.1=py37he98fc37_0
pip=20.2.4=py37h06a4308_0
pixman=0.40.0=h7b6447c_0
pluggy=0.13.1=py37_0
progressbar2=3.37.1=py37_0
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.8=py_0
prompt_toolkit=3.0.8=0
ptyprocess=0.6.0=pyhd3eb1b0_2
py=1.9.0=py_0
py-opencv=3.4.2=py37hb342d67_1
pycparser=2.20=py_2
pygments=2.7.2=pyhd3eb1b0_0
pyopenssl=19.1.0=pyhd3eb1b0_1
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h05f1152_2
pyrsistent=0.17.3=py37h7b6447c_0
pysocks=1.7.1=py37_1
pytest=6.1.1=py37_0
pytest-runner=5.2=py_0
python=3.7.4=h265db76_1
python-dateutil=2.8.1=py_0
python-utils=2.4.0=py37_0
pytorch=1.2.0=cuda100py37h938c94c_0
pytz=2020.1=py_0
pywavelets=1.1.1=py37h7b6447c_2
pyyaml=5.3.1=py37h7b6447c_1
pyzmq=19.0.2=py37he6710b0_1
qt=5.9.7=h5867ecd_1
qtconsole=4.7.7=py_0
qtpy=1.9.0=py_0
readline=7.0=h7b6447c_5
requests=2.22.0=py37_0
scikit-image=0.15.0=py37he6710b0_0
scipy=1.5.2=py37h0b6359f_0
send2trash=1.5.0=py37_0
setuptools=50.3.1=py37h06a4308_1
sip=4.19.8=py37hf484d3e_0
six=1.15.0=py37h06a4308_0
sqlite=3.33.0=h62c20be_0
terminado=0.9.1=py37_0
testpath=0.4.4=py_0
tk=8.6.10=hbc83047_0
toml=0.10.1=py_0
toolz=0.11.1=py_0
torchvision=0.4.0=cuda100py37hecfc37a_0
tornado=6.0.4=py37h7b6447c_1
tqdm=4.32.1=py_0
traitlets=5.0.5=py_0
urllib3=1.24.3=py37_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py37_1
wheel=0.35.1=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py37_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.3=he6710b0_3
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
